python/requirements.txt
stable-baselines3
pyzmq
pandas
schedule
requests
gym
numpy
scikit-learn